fix(collection): vault_progress is System.Double, not i32
Live verification on 2026-04-25 confirmed two field-layout details:
the user's wildcards (1/1/4/12) and gold (1250)/gems (640) all
matched MTGA's UI, but vault_progress read as the sentinel-pattern
0x9999999A. Decoding the 8 bytes at offset 0x70 of
ClientPlayerInventory as a System.Double gives 30.1 — exactly the
percentage MTGA's UI displays.

The 4-byte gap between wcTrackPosition (offset 0x68, i32) and
vaultProgress (offset 0x70) is alignment padding the C# compiler
inserts ahead of the 8-byte-aligned double.

Changes:

- inventory.rs: InventoryValues.vault_progress now f64;
  read_inventory reads 8 bytes for that one field via from_bits;
  test fixture rewired to the new shape.
- chain.rs / run.rs: WalkResult / Snapshot drop their Eq derive
  (PartialEq is enough — f64 doesn't impl Eq).
- lib.rs: WalkSnapshot.vault_progress is f64; the NIF wire shape
  carries the percentage as a float on the Elixir side.
- snapshot.ex: vault_progress field switched to :float plus a
  20260425130000 migration that drop+adds the SQLite column
  (SQLite has no ALTER COLUMN).
- self_check.ex: walker_negative_vault_progress check compares
  against 0.0.
- mem.ex: walker_snapshot type spec marks vault_progress as float().
- Test fixtures across reader_test / self_check_test /
  test_backend_test updated to use 30.1 / 0.0 instead of integers.

Live walker against MTGA pid 2824095 now returns the full snapshot
with all six exposed values matching the in-game UI. 1453 mix
tests + 178 cargo tests pass. The walker is end-to-end verified
against a real MTGA Wine'd process.

+++ lib/scry_2/collection/mem.ex
@@ -77,7 +77,7 @@ defmodule Scry2.Collection.Mem do
           },
           required(:gold) => integer(),
           required(:gems) => integer(),
-          required(:vault_progress) => integer(),
+          required(:vault_progress) => float(),
           required(:build_hint) => String.t() | nil,
           required(:reader_version) => String.t()
         }
